chore(app): Renaming files
