+++ src/app/App.tsx
@@ -1,6 +1,6 @@
-import Header from "@/components/header/Header.tsx";
+import Header from "@/components/Header/Header.tsx";
 import AppRouterProvider from "@/components/router-provider";
-import Footer from "@/components/footer/Footer.tsx";
+import Footer from "@/components/Footer/Footer.tsx";
 import "./app.module.css";
 
 function App() {
