--- src/components/exchange-rate/exchangeRate.module.css
@@ -1,99 +1,0 @@
-.exchangeRate {
-    min-height: 366px;
-    display: flex;
-    justify-content: space-between;
-    padding: 32px;
-    border-radius: 28px;
-    box-shadow: 0 2px 4px 0 rgba(187, 187, 187, 0.2), 0 0 4px 0 rgba(0, 0, 0, 0.08);
-    background: linear-gradient(358.2deg, rgba(178, 178, 178, 0.8) 8.06%, rgba(178, 178, 178, 0) 67.93%),
-    linear-gradient(0deg, rgba(244, 244, 244, 0.9), rgba(244, 244, 244, 0.9));
-}
-
-.exchangeRate__currency {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    gap: 24px;
-}
-
-.exchangeRate__currency h2 {
-    font-family: "Ubuntu", sans-serif;
-    font-size: 30px;
-    font-weight: 500;
-    line-height: 50px;
-    color: rgba(28, 28, 30, 1);
-}
-
-.exchangeRate__text {
-    font-size: 20px;
-    font-weight: 500;
-    line-height: 50px;
-    color: rgba(0, 0, 0, 1);
-}
-
-.currency__block {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 25px;
-    max-width: 440px;
-    justify-content: space-between;
-    padding-top: 8px;
-}
-
-.block__item {
-    display: flex;
-    align-items: center;
-    gap: 20px;
-}
-
-.item__name {
-    font-size: 20px;
-    font-weight: 700;
-    line-height: 22px;
-    letter-spacing: 0.02em;
-    color: rgba(128, 128, 128, 1);
-}
-
-.item__rate {
-    font-size: 20px;
-    font-weight: 700;
-    line-height: 22px;
-    letter-spacing: 0.02em;
-    color: rgba(0, 0, 0, 1);
-}
-
-.exchangeRate__courses {
-    font-size: 20px;
-    font-weight: 700;
-    line-height: 22px;
-    letter-spacing: 0.02em;
-    color: rgba(255, 255, 255, 1);
-    padding-top: 26px;
-}
-
-.exchangeRate__update {
-    display: grid;
-    justify-items: end;
-}
-
-.exchangeRate__update p {
-    font-size: 16px;
-    font-weight: 500;
-    line-height: 50px;
-    color: rgba(0, 0, 0, 1)
-}
-
-@media screen and (max-width: 500px) {
-    .exchangeRate {
-        display: grid;
-    }
-
-    .exchangeRate__update {
-        justify-items: flex-start;
-    }
-
-    .currency__block {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-    }
-}
